requirements.txt
beautifulsoup4
requests
click
pyyaml
